Add CLI and core PDF generation utilities
Provide a command-line interface and extract core PDF generation logic
to support direct CLI usage and reuse from the route handler. This adds
a new bin entry (fumadocs-pdf) in package.json and implements src/cli.ts
to discover pages (via sitemap or explicit paths), generate PDFs, and
emit a manifest. The heavy PDF creation code (puppeteer + pdf-lib) is
moved into src/core.ts and exported from index.ts so the Next
route-handler can call generatePdf/resolveConfig/parseCookies instead of
duplicating logic. The route handler is simplified to use resolveConfig
and generatePdf, and cookie parsing is reused from core. These changes
fix the missing CLI entrypoint error and remove duplicated code while
keeping behavior (expand accordions, lazy image triggering, PDF
outlines) intact.
