+++ packages/pdf-export/package.json
@@ -13,6 +13,9 @@
   "main": "./dist/index.js",
   "module": "./dist/index.js",
   "types": "./dist/index.d.ts",
+  "bin": {
+    "fumadocs-pdf": "./dist/cli.js"
+  },
   "exports": {
     ".": {
       "import": "./dist/index.js",
